requirements.txt
backports.shutil-get-terminal-size==1.0.0
blinker==1.0
decorator==4.0.10
Django==1.7.7
django-filter==0.13.0
django-rest-framework-mongoengine==3.3.0
djangorestframework==3.3.3
ipdb==0.10.1
ipython==4.2.0
ipython-genutils==0.1.0
Markdown==2.6.6
mongoengine==0.9.0
pathlib2==2.1.0
pexpect==4.1.0
pickleshare==0.7.2
ptyprocess==0.5.1
pymongo==2.8
simplegeneric==0.8.1
simplejson==3.8.2
six==1.10.0
traitlets==4.2.1
